requirements.txt
aiobotocore==2.13.0
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
anndata==0.10.8
annotated-types==0.7.0
anthropic==0.29.0
anyio==4.4.0
array_api_compat==1.7.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
biopython==1.83
biothings-client==0.3.1
black==24.4.2
botocore==1.34.106
bs4==0.0.2
cellxgene-census==1.14.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
fuzzywuzzy==0.18.0
gdown==5.2.0
gget==0.28.6
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.2.1
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.81
legacy-api-wrap==1.4
lightning-utilities==0.11.2
littleutils==0.2.2
llvmlite==0.43.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mygene==3.2.2
mypy-extensions==1.0.0
mysql-connector-python==8.0.29
natsort==8.4.0
nest_asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
openai==1.35.3
openpyxl==3.1.4
orjson==3.10.5
outdated==0.2.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
PySocks==1.7.1
PyTDC==0.4.17
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rdkit==2023.9.6
regex==2024.5.15
requests==2.32.3
s3fs==2024.6.0
safetensors==0.4.3
scanpy==1.10.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
session_info==1.0.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
somacore==1.0.11
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
stark_qa==0.1.2
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.12.1
tenacity==8.4.1
threadpoolctl==3.5.0
tiledb==0.29.1
tiledbsoma==1.11.4
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torch_geometric==2.5.3
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.3.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
